No side effects in function arguments
We want to be able to see tests with eqc_gen:sample, we also want to be
able to re-run tests.

Side effects are no good idea if these goals need to be met. Therefore,
we replace the side effect to ask for a port by making a symbolic call
of this.

Nicer solutions are possible, but at least we can now re-run the test
case, and therefore shrink.